compiler: Lift _is_offloadable to DeviceAwareMixin
